OrangeBricks.Web/Controllers/Viewings/Builders/ViewingsOnPropertyViewModelBuilder.cs
using System;
using OrangeBricks.Web.Controllers.Viewings.ViewModels;
using OrangeBricks.Web.Models;
using OrangeBricks.Web.VMBuilder.Interfaces;

namespace OrangeBricks.Web.Controllers.Viewings.Builders
{
    public class ViewingsOnPropertyViewModelBuilder : IViewModelBuilder<ViewinsOnPropertyViewModel, int>
    {
        private readonly IOrangeBricksContext _context;

        public ViewingsOnPropertyViewModelBuilder(IOrangeBricksContext context)
        {
            _context = context;
        }

        public ViewinsOnPropertyViewModel Build(int id)
        {
            throw new NotImplementedException();
        }
    }
}